Move getDuration to the utils folder
