+++ src/components/trip-info.js
@@ -1,16 +1,4 @@
-import {createElement} from "../utils.js";
-
-const getDuration = (startDateUTCTimestamp, endDateUTCTimestamp) => {
-  const startDate = new Date(startDateUTCTimestamp);
-
-  const monthName = startDate.toLocaleString(`en-US`, {
-    month: `short`
-  });
-  const startDay = startDate.getDate();
-  const endDay = new Date(endDateUTCTimestamp).getDate();
-
-  return `${monthName} ${startDay}&nbsp;&mdash;&nbsp;${endDay}`;
-};
+import {createElement, getDuration} from "../utils.js";
 
 export default class TripInfo {
   constructor(cards) {
